+++ functions/ClockSettingsPreview.ps1
@@ -1,21 +1,3 @@
-if ($IsWindows -OR $PSEdition -eq 'desktop') {
-    Try {
-        Add-Type -AssemblyName PresentationFramework -ErrorAction Stop
-        Add-Type -AssemblyName PresentationCore -ErrorAction Stop
-        Add-Type -AssemblyName System.Drawing -ErrorAction Stop
-    }
-    Catch {
-        #Failsafe error handling
-        Throw $_
-        Return
-    }
-}
-else {
-    Write-Warning 'This requires Windows PowerShell or PowerShell Core on Windows.'
-    #Bail out
-    Return
-}
-
 Function Show-PSClockSettingPreview {
     [CmdletBinding()]
     Param()
@@ -47,16 +29,18 @@ Function Show-PSClockSettingPreview {
     }
     Update-TypeData @updateSplat
 
-    Write-Verbose "Starting $($MyInvocation.MyCommand)"
-    Write-Verbose "Running under PowerShell $($PSVersionTable.PSVersion)"
-    Write-Verbose 'Loading font families'
+    _verbose ($strings.Starting -f $MyInvocation.MyCommand)
+    _verbose ($strings.Running -f $PSVersionTable.PSVersion)
+    _verbose ($strings.Detected -f $Host.Name)
+
+    _verbose $strings.LoadingFont
     $Families = [System.Drawing.Text.InstalledFontCollection]::new().Families
-    Write-Verbose 'Loading color values'
+    _verbose $strings.LoadingColor
     $Colors = [System.Drawing.Brushes].GetProperties().name | Select-Object -Skip 1
     $defaultText = $(Get-Date -Format F)
 
-    Write-Verbose "Using default text: $defaultText"
-    Write-Verbose 'Defining the WPF form and controls'
+    _verbose ($strings.UsingText -f $defaultText)
+    _verbose $strings.DefiningWPF
     $window = [System.Windows.Window]@{
         Title                 = $formConfig.Title
         Height                = 400
@@ -232,10 +216,11 @@ Function Show-PSClockSettingPreview {
     })
 
     $window.AddChild($Stack)
-    Write-Verbose 'Showing the form'
+    Write-Information -MessageData $window -Tags wpf
+    _verbose $strings.Showing
     [void]$Window.ShowDialog()
 
-    Write-Verbose "Ending $($MyInvocation.MyCommand)"
-    $global:fc =$formConfig
-    $global:ss = $sliderSize
+    _verbose ($strings.Ending -f $MyInvocation.MyCommand)
+    Write-Information -MessageData $formConfig -tags wpf
+
 }