+++ functions/Get-ScreenWidth.ps1
@@ -1,25 +1,27 @@
 Function Get-PrimaryDisplaySize {
     [cmdletbinding()]
     [OutputType('PSPrimaryDisplaySize')]
-        Param()
+    Param()
 
-        Begin {
-            Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Starting $($MyInvocation.MyCommand)"
-            Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Running under PowerShell version $($PSVersionTable.PSVersion)"
-        } #begin
+    Begin {
+        _verbose ($strings.Starting -f $MyInvocation.MyCommand)
+        _verbose ($strings.Running -f $PSVersionTable.PSVersion)
+        _verbose ($strings.Detected -f $Host.Name)
+    } #begin
 
-        Process {
-            Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Measuring primary display."
-            [PSCustomObject]@{
-                PSTypeName = 'PSPrimaryDisplaySize'
-                Width      = [System.Windows.SystemParameters]::PrimaryScreenWidth
-                Height     = [System.Windows.SystemParameters]::PrimaryScreenHeight
-                WorkArea   = [System.Windows.SystemParameters]::WorkArea
-            }
-        } #process
+    Process {
+        _verbose $strings.Measuring
+        [PSCustomObject]@{
+            PSTypeName = 'PSPrimaryDisplaySize'
+            Width      = [System.Windows.SystemParameters]::PrimaryScreenWidth
+            Height     = [System.Windows.SystemParameters]::PrimaryScreenHeight
+            WorkArea   = [System.Windows.SystemParameters]::WorkArea
+        }
+        Write-Information -MessageData [System.Windows.SystemParameters] -Tags raw
+    } #process
 
-        End {
-            Write-Verbose "[$((Get-Date).TimeOfDay) END    ] Ending $($MyInvocation.MyCommand)"
-        } #end
+    End {
+        _verbose ($strings.Ending -f $MyInvocation.MyCommand)
+    } #end
 
-    } #close Get-ScreenWidth
+} #close Get-ScreenWidth